pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-crm"
version = "0.1.0"
description = "Lightweight CRM backed by Google Sheets with spider-network referral tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Chinchilla AI", email = "engineering@chinchilla.ai"},
]
keywords = ["crm", "google-sheets", "pipeline", "referral-network"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
openclaw-crm = "openclaw_crm.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openclaw_crm"]

[tool.ruff]
line-length = 120
target-version = "py310"
